fix: change requested code
The entire change request are listed in the PR. https://github.com/abhishekpatel946/React-Assignment-10/pull/1
