+++ src/App.scss
@@ -1,5 +1,5 @@
-@import './SCSS/colorStack';
-@import './SCSS/fontStack';
+@import './scss/colorStack';
+@import './scss/fontStack';
 
 .App {
   .App-header {
