+++ src/container/Alert/Alert.jsx
@@ -1,5 +1,5 @@
 import MuiAlert from '@material-ui/lab/Alert';
 
-export function Alert(props) {
+export const Alert = (props) => {
   return <MuiAlert elevation={6} variant='filled' {...props} />;
-}
+};
