+++ src/container/Form/Form-components/style.scss
@@ -1,4 +1,4 @@
-@import '../../../SCSS/colorStack';
+@import '../../../scss/colorStack';
 
 .inputField {
   margin: 10px !important;
