+++ src/container/Table/allTabProps.jsx
@@ -1,4 +1,4 @@
-export function a11yProps(index) {
+export function allTabProps(index) {
   return {
     id: `scrollable-auto-tab-${index}`,
     'aria-controls': `scrollable-auto-tabpanel-${index}`,
