+++ src/container/Table/style.scss
@@ -1,5 +1,5 @@
-@import '../../SCSS/colorStack';
-@import '../../SCSS/fontStack';
+@import '../../scss/colorStack';
+@import '../../scss/fontStack';
 
 MuiBox-root MuiBox-root-89 {
   padding: 0px !important;
